+++ src/controller/settings.py
@@ -166,6 +166,7 @@ class accountSettingsController(globalSettingsController):
   self.dialog.set_value("sound", "indicate_geo", self.config["sound"]["indicate_geo"])
   self.dialog.set_value("sound", "indicate_img", self.config["sound"]["indicate_img"])
   self.dialog.create_extras(OCRSpace.translatable_langs)
+  self.dialog.set_value("extras", "sndup_apiKey", self.config["sound"]["sndup_api_key"])
   self.dialog.realize()
   self.dialog.set_title(_(u"Account settings for %s") % (self.user,))
   self.response = self.dialog.get_response()
@@ -238,9 +239,9 @@ class accountSettingsController(globalSettingsController):
   self.config["sound"]["indicate_audio"] = self.dialog.get_value("sound", "indicate_audio")
   self.config["sound"]["indicate_geo"] = self.dialog.get_value("sound", "indicate_geo")
   self.config["sound"]["indicate_img"] = self.dialog.get_value("sound", "indicate_img")
+  self.config["sound"]["sndup_api_key"] = self.dialog.get_value("extras", "apiKey")
   self.buffer.session.sound.config = self.config["sound"]
   self.buffer.session.sound.check_soundpack()
-
   self.config.write()
 
  def toggle_state(self,*args,**kwargs):