+++ src/extra/AudioUploader/audioUploader.py
@@ -52,7 +52,11 @@ class audioUploader(object):
    self.uploaderDialog = wx_transfer_dialogs.UploadDialog(self.file)
    output.speak(_(u"Attaching..."))
    if self.dialog.get("services") == "SNDUp":
-    url = "http://sndup.net/post.php"
+    base_url = "http://sndup.net/post.php"
+    if len(self.config["sound"]["sndup_api_key"]) > 0:
+     url = base_url + '?apikey=' + self.config['sound']['sndup_api_key']
+    else:
+     url = base_url
    self.uploaderFunction = transfer.Upload(obj=self, field='file', url=url, filename=self.file, completed_callback=completed_callback)
    pub.subscribe(self.uploaderDialog.update, "uploading")
    self.uploaderDialog.get_response(self.uploaderFunction.perform_threaded)