+++ doc/changelog.md
@@ -6,6 +6,7 @@
 * Removed TwUp as service as it no longer exists. ([#112](https://github.com/manuelcortez/TWBlue/issues/112))
 * Release audio files after uploading them. ([#130](https://github.com/manuelcortez/TWBlue/issues/130))
 * Now TWBlue will use Yandex's translation services instead microsoft translator. ([#132](https://github.com/manuelcortez/TWBlue/issues/132))
+* SndUp users will be able to upload audio in their account by using their API Key again. ([#134](https://github.com/manuelcortez/TWBlue/issues/134))
 
 ## Changes in version 0.90